Move import checker to its own file
